# pref(histogram.rs): parallelized histogram generation