### [Solved TODO: check if this can be done in parallel](https://github.com/kornia/kornia-rs/blob/0995adf0b516dfe61a2e98c331af56fd9fc4cf15/crates/kornia-imgproc/src/histogram.rs#L58)

#### Implemented parallel generation of histogram using `rayon` crate optimizing performance as measured below:

1. for kornia-rs/tests/data/apriltags_tag36h11.jpg    
    - without parallelized (RGB)
       - elapsed: 6.328555ms
       - elapsed: 5.98668ms
       - elapsed: 5.968598ms
    - with parallelization (RGB)
       - elapsed: 2.482211ms
       - elapsed: 1.67441ms
       - elapsed: 1.264231ms

2. for kornia-rs/tests/data/dogs.jpeg
    - without parallelized (RGB)
       - elapsed: 749.052µs
       - elapsed: 719.1µs
       - elapsed: 712.953µs
    - with parallelization (RGB)
       - elapsed: 1.045121ms
       - elapsed: 409.218µs
       - elapsed: 413.271µs


#### I measured the elapsed time using:
```rust
let now = std::time::Instant::now();
// logic
let elapsed = now.elapsed();
println!("elapsed: {:?}", elapsed);
```